Test: Move account tests and site test utils to fluxc module
Move account test files from fluxc-tests to fluxc module to
enable internal visibility access for the upcoming Room DAO.
Move SiteTestUtils from fluxc-plugin testFixtures to fluxc
testFixtures since it will need internal DAO access too.
