+++ libs/fluxc/src/test/java/org/wordpress/android/fluxc/store/AccountStoreTest.java
@@ -1,4 +1,4 @@
-package org.wordpress.android.fluxc.account;
+package org.wordpress.android.fluxc.store;
 
 import android.content.Context;
 
@@ -23,7 +23,6 @@ import org.wordpress.android.fluxc.network.rest.wpcom.auth.AccessToken;
 import org.wordpress.android.fluxc.network.rest.wpcom.auth.Authenticator;
 import org.wordpress.android.fluxc.persistence.AccountSqlUtils;
 import org.wordpress.android.fluxc.persistence.WellSqlConfig;
-import org.wordpress.android.fluxc.store.AccountStore;
 import org.wordpress.android.fluxc.store.AccountStore.AuthenticateErrorPayload;
 import org.wordpress.android.fluxc.store.AccountStore.AuthenticatePayload;
 import org.wordpress.android.fluxc.store.AccountStore.AuthenticationErrorType;
